+++ src/protocol/mod.rs
@@ -92,7 +92,7 @@ where
     Ok(result)
 }
 
-#[derive(Deserialize, Debug, Clone)]
+#[derive(Deserialize, Debug, Clone, PartialEq)]
 #[serde(tag = "method")]
 #[allow(clippy::large_enum_variant)]
 pub enum Event {
@@ -120,6 +120,8 @@ pub enum Event {
     ResponseReceived(network::events::ResponseReceivedEvent),
     #[serde(rename = "Log.entryAdded")]
     LogEntryAdded(logs::events::EntryAddedEvent),
+    #[serde(rename = "Runtime.exceptionThrown")]
+    RuntimeExceptionThrown(runtime::events::ExceptionThrownEvent),
 }
 
 #[derive(Deserialize, Debug, Clone)]