+++ tests/extension.rs
@@ -2,12 +2,12 @@ use std::ffi::OsStr;
 
 use failure::Fallible;
 
-use headless_chrome::{browser::default_executable, Browser, LaunchOptionsBuilder};
+use headless_chrome::{browser::default_executable, Browser, LaunchOptions};
 
 #[test]
 fn test_extension() -> Fallible<()> {
     Browser::new(
-        LaunchOptionsBuilder::default()
+        LaunchOptions::default_builder()
             .path(Some(default_executable().unwrap()))
             .extensions(vec![OsStr::new("tests/extension_sampl")])
             .build()