+++ CHANGELOG.md
@@ -2,9 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
-## [Unreleased](https://github.com/atroche/rust-headless-chrome/compare/v0.8.0...HEAD)
+## [Unreleased](https://github.com/atroche/rust-headless-chrome/compare/v0.9.0...HEAD)
 
 ### Added
+* `LaunchOptions::default_builder()` for easier access to LaunchOptionsBuilder (which, because it's created via `derive(Builder)` is hard for editors to deal with
+### Removed
+### Changed
+* Move env_logger to dev dependencies 
+
+## 0.9.0 - 2019-08-22
+
+### Added
+* [Add get_title method to tab](https://github.com/atroche/rust-headless-chrome/pull/167)
+* [Add way to set env vars for Chrome process](https://github.com/atroche/rust-headless-chrome/pull/168)
+* [Improve Runtime Domain, Add ability to listen to all Events on a tab](https://github.com/atroche/rust-headless-chrome/pull/162)
+
 ### Removed
 ### Changed